:tada: typescript support!!!!
