+++ .gitignore
@@ -1,5 +1,6 @@
 /node_modules/
 /public/build/
+.history
 
 .DS_Store
 /.idea/
