+++ package.json
@@ -5,26 +5,27 @@
     "build": "rollup -c",
     "dev": "rollup -c -w",
     "start": "sirv public",
-    "start-server": "deno run -A -L debug --allow-net --allow-read  --allow-write  server.ts"
+    "validate": "svelte-check"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^12.0.0",
-    "@rollup/plugin-node-resolve": "^8.0.0",
     "@types/node": "^14.0.13",
     "node-sass": "^4.14.1",
     "rollup": "^2.3.4",
     "rollup-plugin-livereload": "^1.0.0",
     "rollup-plugin-svelte": "^5.0.3",
     "rollup-plugin-terser": "^5.1.2",
-    "rollup-plugin-typescript2": "^0.27.1",
     "sirv-cli": "^0.4.6",
     "svelte": "^3.0.0",
     "svelte-preprocess": "^3.9.9",
-    "typescript": "^3.9.5"
+    "typescript": "^3.9.5",
+    "@rollup/plugin-node-resolve": "^8.0.0",
+    "@rollup/plugin-commonjs": "13.0.0",
+    "@rollup/plugin-typescript": "^4.0.0"
   },
   "dependencies": {
     "@types/tmi.js": "^1.4.0",
     "localforage": "^1.7.4",
+    "svelte-check": "^0.1.35",
     "tmi.js": "^1.5.0",
     "validate-color": "^2.1.0"
   }
