+++ src/interfaces.ts
@@ -1,0 +1,9 @@
+export interface Cell {
+  color: string;
+  emote?: string;
+}
+
+export interface CellMap {
+  [key: string]: Cell
+}
+
